feat: Added a batch config class with a Job to process the XML files and save them to the database.
Created the class BatchConfig.java with a Job and steps to retrieve xml files as a string list from getAttachments service, map them to MailXmlDocument.java class and call saveDocuments method to save them.
Created the saveDocuments method in DocumentService.java and implemented by DocumentServiceImpl.java to save a list of MailXmlDocument.java.

test: Created tests for BatchConfig step and saveDocuments method.

+++ document-service/src/main/java/com/jesussb/documentservice/document/dto/MailXmlDetail.java
@@ -15,12 +15,12 @@ public class MailXmlDetail {
     @JacksonXmlProperty(localName = "DescItem")
     private String description;
     @JacksonXmlProperty(localName = "QtyItem")
-    private int quantity;
+    private double quantity;
     @JacksonXmlProperty(localName = "UnmdItem")
     private String unit;
     @JacksonXmlProperty(localName = "PrcItem")
-    private int price;
+    private double price;
     @JacksonXmlProperty(localName = "MontoItem")
-    private int amout;
+    private double amout;
 
 }
